Swipe delete, confirm delete, undo ok.

+++ app/src/main/java/com/example/tmdb/fragment/FavoriteFragment.kt
@@ -1,5 +1,7 @@
 package com.example.tmdb.fragment
 
+import android.app.AlertDialog
+import android.graphics.Color
 import android.os.Bundle
 import android.util.Log
 import android.view.LayoutInflater
@@ -9,16 +11,11 @@ import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.RecyclerView
-import androidx.room.Database
 import com.example.tmdb.R
 import com.example.tmdb.RoomTMDBApplication
+import com.example.tmdb.adapter.MenuSeriesAdapterRecyclerView
 import com.example.tmdb.adapter.SearchMovieAdapter
-import com.example.tmdb.data.Movie
-import com.example.tmdb.service.RetrofitInitializer
 import kotlinx.android.synthetic.main.fragment_favorite.*
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
 
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
 private const val ARG_PARAM1 = "param1"
@@ -27,6 +24,7 @@ private const val ARG_PARAM2 = "param2"
 class FavoriteFragment : Fragment() {
 
     private val adapter = SearchMovieAdapter()
+    private val adapterSerie = MenuSeriesAdapterRecyclerView()
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -74,31 +72,63 @@ class FavoriteFragment : Fragment() {
         //------------------------
 
         btRemoveTudo.setOnClickListener {
-            RoomTMDBApplication.movieDao.deleteAll()
+
+            //Alert Dialog to confirm
+            val builder = AlertDialog.Builder(view.context)
+            builder.setTitle("Confirmar remoção")
+            builder.setMessage("Deseja excluir toda a sua lista de favoritos? Tem volta não...")
+            builder.setPositiveButton("SIM") { dialog, which ->
+                Toast.makeText(view.context, "Ok...lista apagada...perdeu heim", Toast.LENGTH_SHORT)
+                    .show()
+
+                RoomTMDBApplication.movieDao.deleteAll()
+                var movies = RoomTMDBApplication.movieDao.getAllMovies() as ArrayList
+                adapter.setList(movies)
+            }
+            builder.setNegativeButton("NÃO PORRA") { dialog, which ->
+                Toast.makeText(view.context, "Ufa...não apaguei viu...", Toast.LENGTH_SHORT).show()
+            }
+
+            builder.setNeutralButton("CANCELA TUDO") { _, _ ->
+                Toast.makeText(view.context, "Cancela a porra toda", Toast.LENGTH_SHORT).show()
+            }
+
+            val dialog: AlertDialog = builder.create()
+
+            dialog.show()
+            dialog.getButton(AlertDialog.BUTTON_NEGATIVE).setTextColor(resources.getColor(R.color.colorPrimary))
+            dialog.getButton(AlertDialog.BUTTON_POSITIVE).setTextColor(Color.GREEN)
+            dialog.getButton(AlertDialog.BUTTON_NEUTRAL).setTextColor(Color.RED)
+
+        }
+
+//        testeSerieBD.setOnClickListener {
+//            var series = RoomTMDBApplication.serieDao.getAllSeries() as ArrayList
+//            adapterSerie.setList(series)
+//            Toast.makeText(context, series[0].toString(), Toast.LENGTH_SHORT).show()
+//            Log.i("SERIES", series[0].toString())
+//        }
+        }
+
+        private fun loadFavoriteDataBase() {
+
             var movies = RoomTMDBApplication.movieDao.getAllMovies() as ArrayList
             adapter.setList(movies)
         }
-    }
 
-    private fun loadFavoriteDataBase() {
+        override fun onResume() {
+            super.onResume()
+            var movies = RoomTMDBApplication.movieDao.getAllMovies() as ArrayList
+            adapter.setList(movies)
+        }
 
-        var movies = RoomTMDBApplication.movieDao.getAllMovies() as ArrayList
-        adapter.setList(movies)
-    }
-
-    override fun onResume() {
-        super.onResume()
-        var movies = RoomTMDBApplication.movieDao.getAllMovies() as ArrayList
-        adapter.setList(movies)
-    }
-
-    companion object {
-        fun newInstance(param1: String, param2: String) =
-            FavoriteFragment().apply {
-                arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
+        companion object {
+            fun newInstance(param1: String, param2: String) =
+                FavoriteFragment().apply {
+                    arguments = Bundle().apply {
+                        putString(ARG_PARAM1, param1)
+                        putString(ARG_PARAM2, param2)
+                    }
                 }
-            }
+        }
     }
-}
